Update to utalize the content-gating middleware
This will allow AB to set a custum content gating function.  In this case it uses a prop set on the site config in order to gate all article content types.

This will also be use along with the base-cms update, https://github.com/parameter1/base-cms/pull/664, to add the ld+json structure in order to crawl and index the gated content.