Adjust auto signup and user check for loginLinkSent hook
Had to remove the !found check on login link sent as it can not promise you will have a omeda encrypted id yet as you could be a new user.  This will attempt to look and check your current subscriptions if an omeda user is found if not just auto sign them up.