Update getting of the omeda encryptedId
Update it to a standalone function that will take the req & payload into consideration.  This is to determine if the current user has an omeda id or if it need to do a rapid identified quick.  This is in order to get that user externalId  to be able to to get its current subscription and help us determine if we need to  sign them up on login link sent.